requirements.txt
# Core dependencies
feedparser==6.0.11
requests==2.31.0
python-dateutil==2.8.2

# AI & Search APIs
groq==0.9.0
tavily-python==0.3.3

# Google Sheets
google-auth==2.29.0
google-api-python-client==2.122.0

# Configuration
PyYAML==6.0.1

# Utilities
python-dotenv==1.0.1
